+++ lib/six_saferpay/models/mastercard_issuer_installments.rb
@@ -4,12 +4,14 @@ module SixSaferpay
     attr_accessor(
       :installment_plans,
       :custom_plan,
+      :chosen_plan,
       :receipt_free_text
     )
 
     def initialize(
         installment_plans: nil,
         custom_plan: nil,
+        chosen_plan: nil,
         receipt_free_text: nil
       )
       if installment_plans
@@ -19,6 +21,10 @@ module SixSaferpay
       if custom_plan
         @custom_plan = SixSaferpay::CustomPlan
           .new(custom_plan)
+      end
+      if chosen_plan
+        @chosen_plan = SixSaferpay::ChosenPlan
+          .new(chosen_plan.to_h)
       end
       @receipt_free_text = receipt_free_text
     end
@@ -30,6 +36,9 @@ module SixSaferpay
       end
       if @custom_plan
         hash.merge!(custom_plan: @custom_plan.to_h)
+      end
+      if @chosen_plan
+        hash.merge!(chosen_plan: @chosen_plan.to_h)
       end
       hash.merge!(receipt_free_text: @receipt_free_text)
       hash
