+++ spec/fabrics/six_saferpay/api/six_transaction/responses/authorize_responses.rb
@@ -5,6 +5,7 @@ SpinningWheel.define do
     payment_means { SpinningWheel.create('response_payment_means') }
     payer { SpinningWheel.create('payer') }
     registration_result { SpinningWheel.create('registration_result') }
+    mastercard_issuer_installments { SpinningWheel.create('mastercard_issuer_installments') }
     liability { SpinningWheel.create('liability') }
     dcc { SpinningWheel.create('dcc') }
   end
