+++ spec/six_saferpay/api/six_transaction/responses/authorize_direct_response_spec.rb
@@ -11,6 +11,7 @@ RSpec.describe SixSaferpay::SixTransaction::AuthorizeDirectResponse do
       payment_means: subject.payment_means.to_h,
       payer: subject.payer.to_h,
       registration_result: subject.registration_result.to_h,
+      mastercard_issuer_installments: subject.mastercard_issuer_installments.to_h
     }
   }
